requirements.txt
# Web框架
fastapi==0.104.1
uvicorn==0.24.0

# HTTP客户端
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3

# RSS生成
feedgen==0.9.0

# 定时任务
apscheduler==3.10.4

# 数据处理
python-dateutil==2.8.2

# 中文分词（用于内容过滤）
jieba==0.42.1

# 环境变量
python-dotenv==1.0.0

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 日志
loguru==0.7.2